Udemy/langchain/intro-to-vector-dbs/pyproject.toml
[tool.poetry]
name = "intro-to-vector-dbs"
version = "0.1.0"
description = "Udemy LangChain vector databases introduction"
authors = ["burmeisterryan3 <ryan.j.burmeister@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
pinecone = "^5.1.0"
python-dotenv = "^1.0.1"
langchainhub = "^0.1.21"
langchain-pinecone = "^0.1.3"
langchain-chroma = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Discover Jupyter notebook files in any specified directories and lint them
extend-include = ["*.ipynb"]

line-length = 79
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Ruff includes 'F' and 'E' codes by default
# Use extend-select to ensure we don't override these settings while adding the below
extend-select = [
    "UP", # pyupgrade
    "D",  # pydocstyle
    "N",  # naming convention
]
ignore = ["D417"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"